+++ src/components/Conversation/index.tsx
@@ -2,7 +2,7 @@ import { Stack, styled } from "@mui/material";
 import useMessageStatus from "../../hooks/Request/useMessageStatus";
 import { selectMessageIds } from "../../store/message/messageSlice";
 import { useAppSelector } from "../../store/store";
-import { LoadingResponseMessage } from "../Chat/LoadingResponseMessage";
+import { LoadingResponseMessage } from "./LoadingResponseMessage";
 import { ScrollButton } from "./ScrollButton";
 import { useScrollToBottom } from "./useScrollToBottom";
 import { FocusableChatChunk } from "../Chat/FocusableChatChunk";
